friday_pjts/pjt09_1/requirements.txt
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
decorator==5.1.1
Django==3.2.9
django-bootstrap-v5==1.0.11
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.11.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
